refactor: flatten single-implementation Security and Mux categories
Rename LibP2P.Security.Noise.* to LibP2P.Noise.* and
LibP2P.Mux.Yamux.* to LibP2P.Yamux.* (tests likewise).

Each category held exactly one implementation, so the intermediate
segment added no information — Noise and Yamux already identify
their stack layer. rust-libp2p takes the same approach, exposing
noise, yamux, and tls as top-level modules with no security/mux
grouping. Future alternatives (TLS, mplex) can sit alongside as
LibP2P.TLS.* / LibP2P.Mplex.*.
